@@ -2,6 +2,6 @@
 class AddMasterAndPositionToFormElements < ActiveRecord::Migration
   def change
     add_column :forms, :master, :boolean, default: false
-    add_column :form_elements, :position, :integer, null:  false, default: 0
+    add_column :form_elements, :position, :integer, null: false, default: 0
   end
 end
