@@ -6,7 +6,7 @@ module ActionKit
     def check_petition_name_is_available(name)
       resp = ActionKit::Client.get('petitionpage', params: { _limit: 1, name: name })
       if resp.code == 200
-        JSON.parse(resp.body)['meta']['total_count'] == 0
+        (JSON.parse(resp.body)['meta']['total_count']).zero?
       else
         false
       end
