@@ -112,7 +112,7 @@ describe Link do
         old_time = Time.now.utc
 
         Timecop.travel(1.hour) do
-          expect { link.update(source: 'BBC') }.to change{
+          expect { link.update(source: 'BBC') }.to change {
             page.reload.updated_at.to_s
           }.from(old_time.to_s).to((old_time + 1.hour).to_s)
         end
